Resolved conflicts in sync/models/models_classification_sync.py and
sync/shariant/query_json_filter.py by taking master's versions.

Reduced the sync hardening down to the Alissa upload response parsing,
which crashed with TypeError when the response omitted the number* keys.
SACGF/variantgrid_private#3831

+++ library/genomics/calculate_cancer_mutation_signatures.py
@@ -179,7 +179,7 @@ class MutationSignatures:
         self.Results = None
 
         NUMPY_DTYPE_TOL = {
-            "float": (np.float, None),
+            "float": (float, None),
             "float64": (np.float64, 1e-15),
             "float128": (np.float128, 1e-30)}
         (self.dtype, self.tolerance) = NUMPY_DTYPE_TOL[precision]
